refactor: refactor batch service and client structure
