--- batch/src/main/java/nexters/dividend/batch/dividend/service/FmpDividendClient.java
@@ -1,11 +1,0 @@
-package nexters.dividend.batch.dividend.service;
-
-/**
- * FMP API 호출 관련 Client 인터페이스입니다.
- *
- * @author Min Ho CHO
- */
-public interface FmpDividendClient {
-
-    void updateDividendData();
-}
